float still result on the same, it might not be because of datatype, it might be algorithm or time constraint

+++ FuelInjectionPerfection/solution.py
@@ -1,6 +1,6 @@
 def solution(n):
     #TODO try reading and implementing bignum python
-    n = int(n)
+    n = float(n)
     count = 0
     while n != 1:
         if n%2==0:
